deps: transformers 5.17.0 (Qwen 3.5), with openpi's patch ported
Qwen 3.5 exists only from transformers 5.2.0, so the 4.53.2 pin cannot
stay; pin 5.17.0 (latest stable). The pin exists for openpi's
`transformers_replace` patch, whose files are verbatim copies of
`transformers` sources, so the two move together: the port and the
pi0.5-side 5.x fixes are in the submodule bump (openpi a385735 on
ryanco/transformers-5, not yet pushed).

egomimic side:
- `from_pretrained(torch_dtype=)` -> `dtype=` in the Qwen3 text stem and
  the two embedding-process scripts.
- `hermetic_env` no longer patches `transformers.utils.hub._is_offline_mode`
  unconditionally: 5.x dropped it and calls `huggingface_hub.is_offline_mode()`,
  which reads the constant we already patch.
- `pi_unavailable` also requires `config.json` in the HF cache: 5.x resolves
  the tokenizer class through the model config, so the tokenizer files alone
  are no longer enough (the cached google/paligemma-3b-mix-224 entry was
  missing it; fetched once online).
- New `tests/unit/test_transformers_upgrade.py`: the installed version
  matches the pyproject pin, `transformers.models.qwen3_5` imports, and
  the openpi patch is actually installed (tuple-returning `GemmaRMSNorm`,
  the AdaRMS config fields, `check()`, `embed_scale == 1`).

CI: the unit-tests job applies openpi's `transformers_replace` overlay after
the install steps -- a stock `uv sync` has none of the patched sources the
pi0.5 port needs, so `test_openpi_patch_*` cannot pass without it -- and the
PaliGemma cache now includes `config.json`, with the key bumped to v2 so the
incomplete v1 entry is not restored instead.

Resolved transitive moves (uv.lock): transformers 4.53.2 -> 5.17.0,
tokenizers 0.21.4 -> 0.23.2, huggingface-hub 0.36.2 -> 1.32.0,
safetensors 0.7.0 -> 0.8.0, hf-xet 1.5.0 -> 1.6.0, click 8.3.3 -> 8.5.0,
and annotated-doc / shellingham / typer added as huggingface-hub 1.x
dependencies. torch, numpy, datasets, timm and lightning are untouched.

The shared `emimic` venv is NOT upgraded by this branch; until it is, run
this branch against /coc/flash7/rco3/EgoVerse-wt/venvs/transformers-upgrade.

+++ egomimic/scripts/embedding_process/qwen3_embedding.py
@@ -102,7 +102,7 @@ class Qwen3TextEmbedding(ZarrKeyTransform):
         self.tokenizer = AutoTokenizer.from_pretrained(
             self.model_name, padding_side="left"
         )
-        self.model = AutoModel.from_pretrained(self.model_name, torch_dtype=self.dtype)
+        self.model = AutoModel.from_pretrained(self.model_name, dtype=self.dtype)
         self.model.to(self.device).eval()
         self._embed_dim = int(self.model.config.hidden_size)
